remove to_bytes/from_bytes methods
We now use Emitable/Parseable traits for NLA values, which is more
consistent with the rest of the structs.

We also removed NetlinkMessage's to_bytes/from_bytes. This makes the API
a bit less user-friendly, but using Emitable/Parseable everywhere is
more consistent. The to_bytes() methods that takes `&mut self` was also
a bit awkward.