fix netlink_proto::Error type
clippy found out that our implementation is recursive:

```
error: using `self` as `Display` in `impl Display` will cause infinite recursion
  --> netlink-proto/src/errors.rs:63:32
   |
63 |             SocketIo(ref e) => write!(f, "{}: {}", self, e),
   |                                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
   = note: `#[deny(clippy::recursive_format_impl)]` on by default
   = help: for further information visit rust-lang.github.io/rust-clippy/master/index.html#recursive_format_impl
   = note: this error originates in the macro `write` (in Nightly builds, run with -Z macro-backtrace for more info)
```

It turns out that this Error implementation is clumsy and that we
don't need the inner `ErrorKind` type. This commit removes the
`ErrorKind` (which is also exposed in `audit` and `rtnetlink`) and
uses `thiserror` to derive most of the boilerplate we need for
`Error`.